b-tcrimer-api/requirements.txt
fastapi
uvicorn
SQLAlchemy
python-dotenv
passlib[bcrypt]
stripe
